make runtime variables
