+++ frontend/common/src/hooks/useLogger.ts
@@ -1,4 +1,8 @@
 /* eslint-disable no-console */
+
+import { SeverityLevel } from "@common/helpers/loggingUtils";
+import { getLoggingLevel } from "@common/helpers/runtimeVariable";
+
 export interface Logger {
   emergency: (message: string) => void;
   alert: (message: string) => void;
@@ -10,42 +14,6 @@ export interface Logger {
   debug: (message: string) => void;
 }
 
-// https://www.rfc-editor.org/rfc/rfc5424
-enum SeverityLevel {
-  Emergency = 0, // system is unusable
-  Alert = 1, // action must be taken immediately
-  Critical = 2, // critical conditions
-  Error = 3, // error conditions
-  Warning = 4, // warning conditions
-  Notice = 5, // normal but significant condition
-  Info = 6, // informational messages
-  Debug = 7, // debug-level messages
-}
-
-// case-insensitive compare
-// https://stackoverflow.com/a/2140723
-function ciEquals(a: string | undefined, b: string | undefined) {
-  return typeof a === "string" && typeof b === "string"
-    ? a.localeCompare(b, undefined, { sensitivity: "accent" }) === 0
-    : a === b;
-}
-
-// parse a string, like from process.env, to the SeverityLevel enum
-const tryParseSeverityLevelString = (
-  s: string | undefined,
-): SeverityLevel | undefined => {
-  if (ciEquals(s, "Emergency")) return SeverityLevel.Emergency;
-  if (ciEquals(s, "Alert")) return SeverityLevel.Alert;
-  if (ciEquals(s, "Critical")) return SeverityLevel.Critical;
-  if (ciEquals(s, "Error")) return SeverityLevel.Error;
-  if (ciEquals(s, "Warning")) return SeverityLevel.Warning;
-  if (ciEquals(s, "Notice")) return SeverityLevel.Notice;
-  if (ciEquals(s, "Info")) return SeverityLevel.Info;
-  if (ciEquals(s, "Informational")) return SeverityLevel.Info;
-  if (ciEquals(s, "Debug")) return SeverityLevel.Debug;
-  return undefined;
-};
-
 // a no-op function for when messages won't be logged
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 const noop = () => {};
@@ -56,9 +24,7 @@ const levelIncludes = (
   messageLevel: SeverityLevel,
 ) => cutoffLevel && messageLevel <= cutoffLevel;
 
-const consoleLoggerLevel = tryParseSeverityLevelString(
-  process.env.LOG_CONSOLE_LEVEL,
-);
+const consoleLoggerLevel = getLoggingLevel("LOG_CONSOLE_LEVEL");
 const consoleLogger: Logger = {
   emergency: levelIncludes(consoleLoggerLevel, SeverityLevel.Emergency)
     ? (message: string) => console.error(message)
@@ -86,9 +52,7 @@ const consoleLogger: Logger = {
     : noop,
 };
 
-const coloredConsoleLoggerLevel = tryParseSeverityLevelString(
-  process.env.LOG_COLORED_CONSOLE_LEVEL,
-);
+const coloredConsoleLoggerLevel = getLoggingLevel("LOG_COLORED_CONSOLE_LEVEL");
 const coloredConsoleStyles = "background: #222; color: #bada55";
 const coloredConsoleLogger: Logger = {
   emergency: levelIncludes(coloredConsoleLoggerLevel, SeverityLevel.Emergency)
