+++ frontend/talentsearch/webpack.base.js
@@ -51,8 +51,6 @@ module.exports = {
         API_SUPPORT_ENDPOINT: JSON.stringify(process.env.API_SUPPORT_ENDPOINT),
         TALENTSEARCH_SUPPORT_EMAIL: JSON.stringify(process.env.TALENTSEARCH_SUPPORT_EMAIL),
         TALENTSEARCH_RECRUITMENT_EMAIL: JSON.stringify(process.env.TALENTSEARCH_RECRUITMENT_EMAIL),
-        LOG_CONSOLE_LEVEL: JSON.stringify(process.env.LOG_CONSOLE_LEVEL),
-        LOG_COLORED_CONSOLE_LEVEL: JSON.stringify(process.env.LOG_COLORED_CONSOLE_LEVEL),
       },
     }),
